+++ progtech/q9nl7w/bea1/src/bea1/Coord.java
@@ -9,26 +9,26 @@ package bea1;
  * @author palyi
  */
 public class Coord {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 
-    public Coord(double x, double y) {
+    public Coord(int x, int y) {
         this.x = x;
         this.y = y;
     }
